Convert billing from integer credits to dollar-based API cost pass-through
Credits now represent real dollar amounts ($10 in credits = $10 of API
usage). Each AI call deducts its actual cost based on per-model pricing
and token counts. Credit packs sell $20/$50 of usage instead of abstract
100/250 credit counts. All credit columns migrated from INTEGER to
NUMERIC(10,6), existing balances reset to zero, and UI updated to show
dollar formatting throughout.

+++ server/app/core/routes/stripe_webhook.py
@@ -68,7 +68,7 @@ async def stripe_webhook(request: Request):
                 try:
                     company_id = UUID(company_id_str)
                     pack = CREDIT_PACKS.get(pack_id)
-                    credits_per_cycle = int(pack["credits"]) if pack else 0
+                    credits_per_cycle = float(pack["credits"]) if pack else 0
                     amount_cents = int(pack["amount_cents"]) if pack else 0
 
                     await billing_service.upsert_subscription(
